+++ sdk/storage/azure-storage-files-datalake/inc/azure/storage/files/datalake/datalake_file_system_client.hpp
@@ -7,7 +7,7 @@
 #include <string>
 
 #include <azure/core/credentials.hpp>
-#include <azure/core/http/pipeline.hpp>
+#include <azure/core/internal/http/pipeline.hpp>
 #include <azure/core/response.hpp>
 #include <azure/storage/blobs/blob_container_client.hpp>
 #include <azure/storage/common/storage_credential.hpp>
@@ -226,12 +226,12 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
   private:
     Azure::Core::Http::Url m_fileSystemUrl;
     Blobs::BlobContainerClient m_blobContainerClient;
-    std::shared_ptr<Azure::Core::Http::HttpPipeline> m_pipeline;
+    std::shared_ptr<Azure::Core::Internal::Http::HttpPipeline> m_pipeline;
 
     explicit DataLakeFileSystemClient(
         Azure::Core::Http::Url fileSystemUrl,
         Blobs::BlobContainerClient blobContainerClient,
-        std::shared_ptr<Azure::Core::Http::HttpPipeline> pipeline)
+        std::shared_ptr<Azure::Core::Internal::Http::HttpPipeline> pipeline)
         : m_fileSystemUrl(std::move(fileSystemUrl)),
           m_blobContainerClient(std::move(blobContainerClient)), m_pipeline(std::move(pipeline))
     {