+++ sdk/storage/azure-storage-files-shares/src/share_client.cpp
@@ -62,7 +62,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     policies.emplace_back(std::make_unique<Storage::Details::SharedKeyPolicy>(credential));
     policies.emplace_back(
         std::make_unique<Azure::Core::Http::TransportPolicy>(options.TransportPolicyOptions));
-    m_pipeline = std::make_shared<Azure::Core::Http::HttpPipeline>(policies);
+    m_pipeline = std::make_shared<Azure::Core::Internal::Http::HttpPipeline>(policies);
   }
 
   ShareClient::ShareClient(const std::string& shareUrl, const ShareClientOptions& options)
@@ -85,7 +85,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     policies.emplace_back(std::make_unique<Storage::Details::StoragePerRetryPolicy>());
     policies.emplace_back(
         std::make_unique<Azure::Core::Http::TransportPolicy>(options.TransportPolicyOptions));
-    m_pipeline = std::make_shared<Azure::Core::Http::HttpPipeline>(policies);
+    m_pipeline = std::make_shared<Azure::Core::Internal::Http::HttpPipeline>(policies);
   }
 
   ShareDirectoryClient ShareClient::GetRootDirectoryClient() const